renamed filenames to stay consistent with new types
