+++ server/src/actions/deleteConnectedUser.ts
