+++ server/src/actions/getConnectedUsers.ts
