+++ server/src/actions/updateConnectedUser.ts
